+++ src/main/java/database/services/ProglangService.java
@@ -16,7 +16,7 @@ public class ProglangService {
     public ProglangService() {
     }
 
-    public Proglang findProglang(Integer proglang_id) {
+    public Proglang findProglang(long proglang_id) {
         return proglangDao.findByProglangId(proglang_id);
     }
 
@@ -26,7 +26,7 @@ public class ProglangService {
      * @param proglang_id id языка программирования
      * @return количество вопросов по выбранному ЯП
      */
-    public Long getSizeOfProglang(Integer proglang_id) {
+    public long getSizeOfProglang(long proglang_id) {
         return proglangDao.countProgquizzesByProglangId(proglang_id);
     }
 
@@ -34,10 +34,9 @@ public class ProglangService {
      * Получение конкретного вопроса по ЯП
      * @param proglang_id id ЯП
      * @param offset отступ(начиная с 0, может быть null)
-     * @param limit сколько вопросов(может быть null)
      * @return вопрос
      */
-    public Progquiz getQuestionByLang(Integer proglang_id, int offset) {
+    public Progquiz getQuestionByLang(long proglang_id, int offset) {
         return proglangDao.findProgquizByProglangId(proglang_id, offset);
     }
 
@@ -46,7 +45,7 @@ public class ProglangService {
      * @param proglang_name имя ЯП
      * @return id ЯП, либо -1 если ЯП не существует
      */
-    public int getProglangIdByName(String proglang_name) {
+    public long getProglangIdByName(String proglang_name) {
         return proglangDao.getIdByName(proglang_name);
     }
